+++ examples/src/todomvc-custom-elements/src/main.ts
@@ -2,7 +2,7 @@
 // import 'todomvc-app-css/index.css'
 // import './index.css'
 
-import { createStorePromise, querySQL, rowQuery } from '@livestore/livestore'
+import { createStorePromise, query, rowQuery } from '@livestore/livestore'
 import { cuid } from '@livestore/utils/cuid'
 import { makeAdapter } from '@livestore/web'
 import LiveStoreSharedWorker from '@livestore/web/shared-worker?sharedworker'
@@ -24,7 +24,7 @@ const store = await createStorePromise({
 })
 
 const appState$ = rowQuery(tables.app)
-const todos$ = querySQL(`select * from todos where deleted is null`, { schema: Schema.Array(tables.todos.schema) })
+const todos$ = query(`select * from todos where deleted is null`, { schema: Schema.Array(tables.todos.schema) })
 
 const updateNewTodoText = (text: string) => store.mutate(mutations.updateNewTodoText({ text }))