+++ examples/src/todomvc/src/components/MainSection.tsx
@@ -1,4 +1,4 @@
-import { querySQL, rowQuery, SessionIdSymbol, sql } from '@livestore/livestore'
+import { query, rowQuery, SessionIdSymbol, sql } from '@livestore/livestore'
 import { useQuery, useStore } from '@livestore/react'
 import { Schema } from 'effect'
 import React from 'react'
@@ -19,7 +19,7 @@ const filterClause$ = rowQuery(tables.app, SessionIdSymbol, {
 // We create a new reactive SQL query which interpolates the filterClause.
 // Notice how we call filterClause() as a function--
 // that gets the latest value of that reactive query.
-const visibleTodos$ = querySQL((get) => sql`select * from todos ${get(filterClause$)}`, {
+const visibleTodos$ = query((get) => sql`select * from todos ${get(filterClause$)}`, {
   schema: Schema.Array(tables.todos.schema),
   label: 'visibleTodos',
 })