+++ packages/@livestore/react/src/useQuery.test.tsx
@@ -1,4 +1,4 @@
-import { querySQL } from '@livestore/livestore'
+import { query } from '@livestore/livestore'
 import { Effect, Schema } from '@livestore/utils/effect'
 import { renderHook } from '@testing-library/react'
 import React from 'react'
@@ -14,7 +14,7 @@ describe('useQuery', () => {
 
       const renderCount = makeRenderCount()
 
-      const allTodos$ = querySQL(`select * from todos`, { schema: Schema.Array(tables.todos.schema) })
+      const allTodos$ = query(`select * from todos`, { schema: Schema.Array(tables.todos.schema) })
 
       const { result } = renderHook(
         () => {
@@ -41,11 +41,11 @@ describe('useQuery', () => {
 
       const renderCount = makeRenderCount()
 
-      const todo1$ = querySQL(`select * from todos where id = 't1'`, {
+      const todo1$ = query(`select * from todos where id = 't1'`, {
         label: 'libraryTracksView1',
         schema: Schema.Array(tables.todos.schema),
       })
-      const todo2$ = querySQL(`select * from todos where id = 't2'`, {
+      const todo2$ = query(`select * from todos where id = 't2'`, {
         label: 'libraryTracksView2',
         schema: Schema.Array(tables.todos.schema),
       })